refactoring DAO
