+++ app/src/main/java/com/skydoves/githubfollows/room/HistoryDao.kt
@@ -5,6 +5,7 @@ import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Insert
 import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.Query
+import com.skydoves.githubfollows.models.History
 
 /**
  * Developed by skydoves on 2018-01-27.
